refactored utils in weak supervision package
