+++ dpd/weak_supervision/context_window_functions/bag_window_function.py
@@ -21,7 +21,7 @@ from dpd.models.embedder import CachedTextFieldEmbedder
 from dpd.utils import TensorList
 from dpd.weak_supervision.feature_extractor import FeatureExtractor, FeatureCollator
 
-from .utils import get_context_window, get_context_range, label_index, NEGATIVE_LABEL
+from ..utils import get_context_window, get_context_range, label_index, NEGATIVE_LABEL
 from .window_function import WindowFunction
 
 class BagWindowFunction(WindowFunction):
