+++ dpd/weak_supervision/contextual_functions/cwr_knn.py
@@ -21,7 +21,7 @@ from dpd.weak_supervision import WeakFunction, AnnotatedDataType, AnnotationType
 from dpd.models.embedder import CachedTextFieldEmbedder
 from dpd.utils import TensorList
 
-from .utils import get_label_index, construct_train_data, extract_features, NEGATIVE_LABEL
+from ..utils import get_label_index, construct_train_data, extract_features, NEGATIVE_LABEL
 
 class CWRkNN(WeakFunction):
     '''
