+++ dpd/weak_supervision/contextual_functions/cwr_linear.py
@@ -20,7 +20,7 @@ from dpd.models.embedder import CachedTextFieldEmbedder
 from dpd.models import construct_linear_classifier, LinearType
 from dpd.utils import TensorList
 
-from .utils import get_label_index, construct_train_data, extract_features, NEGATIVE_LABEL
+from ..utils import get_label_index, construct_train_data, extract_features, NEGATIVE_LABEL
 
 class CWRLinear(WeakFunction):
     def __init__(
